pyproject.toml
[project]
name = "inkwell"
version = "0.1.0"
description = "Modern web fiction to EPUB downloader"
readme = "README.md"
authors = [
    { name = "Oswin", email = "oswin@skyworkz.nl" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx[http2]",
    "beautifulsoup4",
    "lxml",
    "typer[all]",
    "rich",
    "pydantic>=2.0",
    "pydantic-settings",
    "ebooklib",
    "Pillow",
    "loguru",
    "tenacity",
    "platformdirs",
    "anyio",
]

[project.scripts]
inkwell = "inkwell.cli.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/inkwell"]
